% Generated by roxygen2 (4.0.1): do not edit by hand
\name{couchdb_server}
\alias{couchdb_server}
\title{Configure the location of the CouchDB sever}
\usage{
couchdb_server(uri)
}
\arguments{
\item{uri}{If missing, then the current URI is returned. Otherwise
  the CouchDB URI set to the given one.}
}
\description{
Configure the location of the CouchDB sever
}

