% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cran_mirror}
\alias{cran_mirror}
\title{Configure the location of the CRAN mirror}
\usage{
cran_mirror(path)
}
\arguments{
\item{path}{If not missing, then the path to the CRAN mirror.
  If missing, then the current setting is returned.}
}
\description{
Configure the location of the CRAN mirror
}

