% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cran_releases}
\alias{cran_releases}
\title{Package versions snapshotted for R releases}
\usage{
cran_releases(version, format = c("version", "short", "full"))
}
\arguments{
\item{version}{R version number.}

\item{format}{What to return. \sQuote{\code{version}} means only the
  version numbers that were current at the given R
  release. \sQuote{\code{short}} also includes the title of the packages}
}
\value{
List of packages.
}
\description{
Package versions snapshotted for R releases
}

