% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/couch.r
\name{create_empty_db}
\alias{create_empty_db}
\title{Create an empty Couchdb database with CRAN-DB structure}
\usage{
create_empty_db()
}
\description{
Create an empty Couchdb database with CRAN-DB structure
}

