% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing.r
\name{scrape_metacran_version}
\alias{scrape_metacran_version}
\title{Scrape Metacran versions}
\usage{
scrape_metacran_version()
}
\description{
Helper function to scrape the metacran version for all CRAN packages, in order
to determine packages that are missing or outdated on metacran.
}
